refactor: NewsRepository -> extends JpaRepository. package dao -> repository
