--- news-service/src/main/java/dao/NewsRepository.java
@@ -1,9 +1,0 @@
-package dao;
-
-import org.springframework.data.repository.PagingAndSortingRepository;
-import org.springframework.stereotype.Repository;
-import ru.clevertec.newsservice.entity.News;
-
-@Repository
-public interface NewsRepository extends PagingAndSortingRepository<News, Long> {
-}
