+++ news-service/src/main/java/ru/clevertec/NewsServiceApplication.java
@@ -1,3 +1,5 @@
+package ru.clevertec;
+
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
