+++ news-service/src/main/java/ru/clevertec/entity/BaseEntity.java
@@ -1,4 +1,4 @@
-package entity;
+package ru.clevertec.entity;
 
 import java.io.Serializable;
