+++ news-service/src/main/java/ru/clevertec/model/Role.java
@@ -1,4 +1,4 @@
-package model;
+package ru.clevertec.model;
 
 public enum Role {
