+++ news-service/src/main/java/ru/clevertec/repository/CommentRepository.java
@@ -1,8 +1,8 @@
-package dao;
+package ru.clevertec.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import ru.clevertec.newsservice.entity.Comment;
+import ru.clevertec.entity.Comment;
 
 @Repository
 public interface CommentRepository extends JpaRepository<Comment, Long> {
