+++ news-service/src/main/java/ru/clevertec/repository/NewsRepository.java
@@ -1,0 +1,9 @@
+package ru.clevertec.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+import ru.clevertec.entity.News;
+
+@Repository
+public interface NewsRepository extends JpaRepository<News, Long> {
+}
